Delete

+++ index.php
@@ -26,7 +26,7 @@ if ($statement->rowCount() > 0) {
             <td><?= $usuario['email']; ?></td>
             <td>
                 <a href="editar.php?id=<?=$usuario['id']; ?>">[ Editar ]</a>
-                <a href="excluir.php?id=<?=$usuario['id']; ?>">[ Excluir ]</a>
+                <a href="excluir.php?id=<?=$usuario['id']; ?>" onclick="return confirm('tem certeza que deseja excluir ?')">[ Excluir ]</a>
             </td>
         </tr>
     <?php endforeach; ?>
